requirements.txt
# LLM & NLP
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
openai>=1.50.0

# Vector Database
chromadb>=0.5.0

# Embeddings & Reranking
sentence-transformers>=3.0.0

# Ontology & Knowledge Graph
rdflib>=7.0.0
networkx>=3.2
pyvis>=0.3.2

# Dashboard
streamlit>=1.38.0
plotly>=5.22.0

# Document Processing
python-docx>=1.1.0
pypdf>=3.17.0

# Evaluation
ragas>=0.1.0
datasets>=2.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
